requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
editor==1.6.6
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.0.2
opencv-python==4.10.0.84
pillow==10.4.0
proglog==0.1.10
python-dotenv==1.0.1
requests==2.32.3
runs==1.2.2
tqdm==4.67.1
urllib3==2.2.3
xmod==1.8.1
